23.09.29 -annual 연차 시작일,종료일 수정 -연차 승인 대기중일때만 삭제 가능하도록 구현 -연차 승인 대기중일때 수정할 수 있도록 구현중

+++ src/main/java/com/project/ownote/annual/repository/AnnualEm.java
@@ -1,7 +1,6 @@
 package com.project.ownote.annual.repository;
 
-import com.project.ownote.annual.dto.Annual;
-import com.project.ownote.annual.dto.AnnualPage;
+import com.project.ownote.annual.entity.Annual;
 import org.springframework.stereotype.Repository;
 
 import javax.persistence.EntityManager;
@@ -16,11 +15,12 @@ public class AnnualEm {
 
 
     @Transactional
-    public void annDeleteById(Long annual_id){
-        Annual annual = em.find(Annual.class,annual_id);
-        em.remove(annual);
+    public void annDeleteById(Long annual_id) {
+        Annual annual = em.find(Annual.class, annual_id);
+        if (annual != null) {
+            em.remove(annual);
+        }
     }
-
 
     public Annual annFindById(Long annual_id){
         Annual annual = em.find(Annual.class,annual_id);
@@ -29,8 +29,8 @@ public class AnnualEm {
 
 
     @Transactional
-    public void annModifyById(Long annual_id){
-        em.merge(annual_id);
+    public void annModify(Annual annual){
+        em.merge(annual);
 
     }
